Modifed constructors to make GCC compilation work
Inheriting the constructors in the movetype class was causing
compilation with GCC to fail. Also explicilty deleted the copy
constructors and assigners on the movetype, system bias, and system
order param classes; this may not have been necessary but it actually
provides some confidence that all the explicit calls to the
constructors I've had to put in because of the use of virtual base
classes is not accidently making two copies of a base class.