Major reogranization of python modules and scripts
The last remnants of the simulation aspects of the python module have
been removed. The module has also been renamed and is now the main
place that python code should be placed. The modules are currently in
a non-working state. As it comes time to use each piece I will fix it
then. I have python scripts in the scripts directory, but these should
mostly be thought of as example interfaces to the main analsyis modules and
should mainly be parsing and sending out command line arguments to the
functions, which may be modified on a per calculation basis as neeed.