Fixed RG bug; added new OP; settable seed
There was a major mistake in the RG implementation that effected the
way the number of available configs was calculated for both the old and
new configs. A new order parameter that measures whether two domains
are on adjacent lattice sites was added. Random seeds can now be set in
the input file.