Removed extra external bias updates in CB
I only need to calculate it once, for the new config, which will have
the difference between the new and old. I don't need to update after
unassigning and then again after reassigning, and both again for the
old config. The biases are always called either after growing new or
after rejecting and resetting.