Stacking and volume bugfixes, new potential
The linear helix checks had a bug relating to the check for linear
helices that require checking across strands. It would skip checks
because of a bad call to checking for doubly contiguous helices. I also
made a mistake in the last commit on the change to the volume
calculations from molarity when removing the factor of 4. Finally, I
have added a new potential that derives from the linear helix one that
adds additional constraints to kinks.