Remove US production and add interation restart
There was no meaniful difference between iterations and the production,
so I removed the production stuff and have iterations output the weights
they use and the ones they calculate. I added archiving of the internal
variables necessary to restart in the middle of an interation. I may
need to play with restart steps and duration, but for now it will just
assume it has to complete whatever total duration is given in the
restart input file.