Add reading and writing of US variables
The internal variables of US simulations are now written to file in a
boost serialized archive and are able to read back in to restart in the
middle of an iteration. I also merged iterations and the production as
there was no meaniful difference. I have added boolean flags for the
parameter files to specify whether something is being read from file but
I have only incorporated their use into the US code.