Use restart iter steps and remove window debugging
When restarting from an iteration in PTMWUS, update the starting step to
be the total number of steps completed in the iteration being restarted
from. This is necessary to calculate the weights at the end of the
iteration correcty. Also added a check that the weights add up to 1.
Commented out the debugging messages for now as they are creating
oversized output files on the cluster on multi day jobs.