Improve error handling
I have updated the custom exceptions to be derived from the standard
library runtime_error exception class, and now check all files for
existance before trying to open them. I have wrapped the contents of
main in a try catch statment and print out the contents any standard
library derived exception there. Better messages are now included in
many of the file reading to say which file has the problem and whether
it just isn't there, or whether there is a problem with how its
formatted.